requirements.txt
selenium >= 4.6.0
webdriver-manager >= 3.8.5
python-dotenv >= 0.21.0
undetected-chromedriver >= 3.1.7
